Adding basic HTTP support to be able to query Viessman interface using regular http calls in addition to the provided telnet interface.
An simple basic HTTP server is started at the same time as the socket server.

All actions registered in the config files are registered to /get_<thing_id>_<channel_id>

It will return the result regardless of HTTP Verb.

The HTTP server does not yet provide a method to write data to the interface.